requirements.txt
############################
# Core API (FastAPI / ASGI)
############################
fastapi==0.128.0
starlette==0.49.3
uvicorn==0.39.0
anyio==4.12.1
h11==0.16.0
sniffio==1.3.1
exceptiongroup==1.3.1
typing_extensions==4.15.0
idna==3.11
certifi==2026.1.4
urllib3==1.26.20
packaging==24.2

############################
# HTTP Client (external API calls)
############################
httpx==0.28.1
httpcore==1.0.9

############################
# Background Tasks (Celery + Redis)
############################
celery==5.6.2
redis==7.0.1
amqp==5.3.1
billiard==4.2.4
kombu==5.6.2
vine==5.1.0
async-timeout==5.0.1

############################
# CLI / REPL helpers (pulled in by Celery tooling)
############################
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.2.14

############################
# AWS / Object Storage (S3 + MinIO compatible)
############################
boto3==1.42.22
botocore==1.42.22
s3transfer==0.16.0
jmespath==1.0.1

############################
# Database (Postgres + ORM + pgvector)
############################
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
pgvector==0.4.2
alembic==1.13.1

############################
# LLM + RAG Orchestration (LangChain + OpenAI)
############################
langchain==1.2.3
langchain-core==1.2.6
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
openai>=1.0.0
tiktoken==0.8.0
tenacity==9.0.0
tqdm==4.67.1
distro==1.9.0
jiter==0.12.0

############################
# Validation / Settings (Pydantic)
############################
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
annotated-types==0.7.0
typing-inspection==0.4.2
annotated-doc==0.0.4

############################
# Auth (password hashing + JWT tokens)
############################
passlib==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0

############################
# Document parsing / ingestion
############################
pypdf==3.1.0
docx2txt==0.9

############################
# Environment variables & Time utilities
############################
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
